i feel sonic is good

+++ 1.4 sonic/sonic.h
@@ -4,6 +4,7 @@
 #include "boot.h"
 
 void sonic_send();
-uint sonic_measure();
+uchar sonic_measure();
+
 
 #endif
